fix db

+++ events/messageReactionAdd.js
@@ -4,20 +4,15 @@ const createRow = require("../functions/messages/createRow");
 const sendEmbed = require("../functions/messages/sendEmbed");
 
 client.on(Events.MessageReactionAdd, async (messageReaction, user) => {
-  // console.log(user);
   if (user.bot) return;
 
   const { message, emoji, users } = messageReaction;
-
-  // console.log("reakcja:", messageReaction);
 
   const member = await message.guild.members.fetch(user.id); // Pobierz informacje o członku
 
   if (member.permissions.has(PermissionsBitField.Flags.Administrator)) {
     if (message.channel.name === "ADD IMAGES")
-      if (emoji.name === "❌")
-        // console.log(message.guild.members.cacher);
-        message.delete();
+      if (emoji.name === "❌") message.delete();
   } else {
     if (message.channel.name.includes("listing-")) {
       console.log(`Collected ${emoji.name} from ${user.tag}`);
@@ -28,7 +23,6 @@ client.on(Events.MessageReactionAdd, async (messageReaction, user) => {
         const messages = await message.channel.messages.fetch();
         if (
           !messages.some(m => {
-            // console.log(m);
             return m.embeds[0]?.description.includes(
               "**, you can't add filters.\n\nOnly **admins** have permission to do so."
             );
@@ -43,13 +37,11 @@ client.on(Events.MessageReactionAdd, async (messageReaction, user) => {
           `Usunięto reakcję ${emoji.name} od użytkownika ${user.tag}, który nie jest administratorem.`
         );
       } catch (error) {
-        console.error("\x1b[31m%s\x1b[0m", `Wystąpił błąd podczas usuwania reakcji: ${error}`);
+        console.error(
+          "\x1b[31m%s\x1b[0m",
+          `Wystąpił błąd podczas usuwania reakcji: ${error}`
+        );
       }
     }
   }
-
-  // if (member.permissions.has(PermissionsBitField.Flags.Administrator))
-  // if (message.channel.name === "ADD IMAGES")
-  // // console.log(message.guild.members.cacher);
-  // if (emoji.name === "❌") message.delete();
 });
